refactor: Update user client hooks to include user.Hooks
