+++ internal/auth/ent/client.go
@@ -658,7 +658,8 @@ func (c *UserClient) QueryTokens(u *User) *TokenQuery {
 
 // Hooks returns the client hooks.
 func (c *UserClient) Hooks() []Hook {
-	return c.hooks.User
+	hooks := c.hooks.User
+	return append(hooks[:len(hooks):len(hooks)], user.Hooks[:]...)
 }
 
 // Interceptors returns the client interceptors.
