+++ internal/auth/ent/runtime.go
@@ -2,68 +2,4 @@
 
 package ent
 
-import (
-	"auth/ent/role"
-	"auth/ent/schema"
-	"auth/ent/token"
-	"auth/ent/user"
-	"time"
-)
-
-// The init function reads all schema descriptors with runtime code
-// (default values, validators, hooks and policies) and stitches it
-// to their package variables.
-func init() {
-	roleFields := schema.Role{}.Fields()
-	_ = roleFields
-	// roleDescCreatedAt is the schema descriptor for created_at field.
-	roleDescCreatedAt := roleFields[3].Descriptor()
-	// role.DefaultCreatedAt holds the default value on creation for the created_at field.
-	role.DefaultCreatedAt = roleDescCreatedAt.Default.(func() time.Time)
-	// roleDescUpdatedAt is the schema descriptor for updated_at field.
-	roleDescUpdatedAt := roleFields[4].Descriptor()
-	// role.DefaultUpdatedAt holds the default value on creation for the updated_at field.
-	role.DefaultUpdatedAt = roleDescUpdatedAt.Default.(func() time.Time)
-	// role.UpdateDefaultUpdatedAt holds the default value on update for the updated_at field.
-	role.UpdateDefaultUpdatedAt = roleDescUpdatedAt.UpdateDefault.(func() time.Time)
-	// roleDescID is the schema descriptor for id field.
-	roleDescID := roleFields[0].Descriptor()
-	// role.DefaultID holds the default value on creation for the id field.
-	role.DefaultID = roleDescID.Default.(func() string)
-	tokenFields := schema.Token{}.Fields()
-	_ = tokenFields
-	// tokenDescRevoked is the schema descriptor for revoked field.
-	tokenDescRevoked := tokenFields[4].Descriptor()
-	// token.DefaultRevoked holds the default value on creation for the revoked field.
-	token.DefaultRevoked = tokenDescRevoked.Default.(bool)
-	// tokenDescCreatedAt is the schema descriptor for created_at field.
-	tokenDescCreatedAt := tokenFields[5].Descriptor()
-	// token.DefaultCreatedAt holds the default value on creation for the created_at field.
-	token.DefaultCreatedAt = tokenDescCreatedAt.Default.(func() time.Time)
-	// tokenDescUpdatedAt is the schema descriptor for updated_at field.
-	tokenDescUpdatedAt := tokenFields[6].Descriptor()
-	// token.DefaultUpdatedAt holds the default value on creation for the updated_at field.
-	token.DefaultUpdatedAt = tokenDescUpdatedAt.Default.(func() time.Time)
-	// token.UpdateDefaultUpdatedAt holds the default value on update for the updated_at field.
-	token.UpdateDefaultUpdatedAt = tokenDescUpdatedAt.UpdateDefault.(func() time.Time)
-	// tokenDescID is the schema descriptor for id field.
-	tokenDescID := tokenFields[0].Descriptor()
-	// token.DefaultID holds the default value on creation for the id field.
-	token.DefaultID = tokenDescID.Default.(func() string)
-	userFields := schema.User{}.Fields()
-	_ = userFields
-	// userDescCreatedAt is the schema descriptor for created_at field.
-	userDescCreatedAt := userFields[4].Descriptor()
-	// user.DefaultCreatedAt holds the default value on creation for the created_at field.
-	user.DefaultCreatedAt = userDescCreatedAt.Default.(func() time.Time)
-	// userDescUpdatedAt is the schema descriptor for updated_at field.
-	userDescUpdatedAt := userFields[5].Descriptor()
-	// user.DefaultUpdatedAt holds the default value on creation for the updated_at field.
-	user.DefaultUpdatedAt = userDescUpdatedAt.Default.(func() time.Time)
-	// user.UpdateDefaultUpdatedAt holds the default value on update for the updated_at field.
-	user.UpdateDefaultUpdatedAt = userDescUpdatedAt.UpdateDefault.(func() time.Time)
-	// userDescID is the schema descriptor for id field.
-	userDescID := userFields[0].Descriptor()
-	// user.DefaultID holds the default value on creation for the id field.
-	user.DefaultID = userDescID.Default.(func() string)
-}
+// The schema-stitching logic is generated in auth/ent/runtime/runtime.go
